Specify text-mode for files and pipes explicitly
While in python2 it was possible to implicitly convert between binary
data ('bytes' objects) and strings ('str'), python3 enforces a strict
distinction of these types. It is necessary explicitly convert binary
data to text by specification of the encoding.

Using text-mode on file-object level allows to take advantage of the
default locale and builtin capabilities.