Include and use pandoc in wheel for OS X
I copied a Homebrew-built pandoc and libgmp, using delocate to rewrite
the pandoc binary to look for the library in ../lib instead of
/usr/local/lib.

This pandoc binary and libgmp dylib get included in wheels.

If we can't find `pandoc` and `sys.platform == darwin` then we use the
darwin pandoc in the wheel.