Check pandoc return code after conversation
Until now, a wrong commandline parameter and therfore a pandoc
error without a conversation was not signaled to the caller
and an empty string was returned. The problem was only seen
on the commandline, were the error was printed to stderr.

Now check the return code and return any error message in the
exception message.