Better default locations for downloads and for searching for pandoc binary
The default download path is now set to a platform specific path:

* ~/bin on Linux
* ~\AppData\Local\Pandoc on windows (also the default install location for pandoc)
* ~/Applications/pandoc on Mac OS X

In addition to these paths we also look into ~/.bin on linux and the python
script/bin path. The last ensures that pandoc can be installed into a virtual
env and is found even if the env is not activated (only run a python directly
from that path).

Together these changes ensure that

* conda packages can be created for pandoc and these pandoc installs are found.
* even if a default pandoc install wasn't added to the PATH on windows, it will
  be found.
* On all platforms, we have a proper default install location if a user will use
  `pypandoc.pandoc_download.download_pandoc()`