Add verify_formats option
This commit adds the verify_formats boolean flag to all convert_*
functions. This allows the caller to directly call pandoc rather than
having pypandoc run three subprocess calls before calling pandoc to
convert the text. This flag defaults to True, and should only be set
False when confident of the formats *and* performance is an issue.

For example, this was originally created to speed up compilation of a
document which consisted of over 10,000 html files to be converted to
latex.