Tests fixes (#328)
* tests: fix invocation with sys.executable

The formatting to substitute sys.executable was not effective because the
formatted string was discarded. But even when when the substitution is made,
pandoc ignores the shebang line in the file when the file is not executable and
invokes it using 'python'. This second issue was hiding the first. We need to
fix both to actually execute code using the desired sys.executable. Finally,
the first .format() was interfering with the second .format().

* tests: fix indentation

* tests: show diff if comparison fails